add-splash-time
1.add function in model qemu: splash time before booting, menu on/off, splash picture
2.add an infrasim logo as splash time
3.add a test for splash time
4.modify the boot-order test with timeout to save test time
